changed eeprom to 32 bits to reserve blocks for future features

+++ tmk_core/common/eeconfig.h
@@ -41,7 +41,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define EECONFIG_KEYBOARD                          (uint32_t *)15
 #define EECONFIG_USER                              (uint32_t *)19
 
-#define EECONFIG_HAPTIC                            (uint16_t*)24
+#define EECONFIG_HAPTIC                            (uint32_t*)24
 
 /* debug bit */
 #define EECONFIG_DEBUG_ENABLE                       (1<<0)
@@ -97,8 +97,8 @@ uint32_t eeconfig_read_user(void);
 void eeconfig_update_user(uint32_t val);
 
 #ifdef HAPTIC_ENABLE
-uint16_t eeconfig_read_haptic(void);
-void eeconfig_update_haptic(uint16_t val);
+uint32_t eeconfig_read_haptic(void);
+void eeconfig_update_haptic(uint32_t val);
 #endif
 
 #endif
